# Remove unnecessary usage of _TSObject