This is part 2 in an N-part series of PRs to disentangle inter-dependent pieces of tslib.pyx (and by extension, lib.pyx and period.pyx).

`tslib` has a `_TSObject` class that is used as a container during conversion steps.  In a number of the places where it is currently used, it is not needed.  All this PR does is remove it in cases where it is either unused or unneeded.

- [ ] closes #xxxx
- [ ] tests added / passed
- [ ] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry