@jreback 2017-08-20
pls show asv for all time series operations