@jbrockmendel 2017-08-20
> pls show asv for all time series operations

I'll get that started shortly.