@jbrockmendel 2017-08-21
OK.  Let's start with `to_pydatetime`.

```
In [3]: ts = pd.Timestamp("2017-09-01 09:00:00")
In [4]: %timeit ts.to_pydatetime()
```
In master this came back with 
```
The slowest run took 20.69 times longer than the fastest. This could mean that an intermediate result is being cached.
1000000 loops, best of 3: 484 ns per loop
```
Under the PR this was in fact much slower:
```
In [4]: %timeit ts.to_pydatetime()
The slowest run took 109.78 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 5.95 µs per loop
```

So I reverted the edits to `py_datetime`.  Post-reversion that measurement back to parity (well,1.03% slower).
```
In [21]: %timeit ts.to_pydatetime()
The slowest run took 24.39 times longer than the fastest. This could mean that an intermediate result is being cached.
1000000 loops, best of 3: 489 ns per loop
```

All subsequent timings here are post-reversion.
```
In [26]: s = pd.date_range("2016-01-01", periods=10000, freq='H')
In [27]: sb = pd.date_range("2016-01-01", periods=10000, freq='BH')
In [28]: %timeit s.to_pydatetime()
# PR     --> 100 loops, best of 3: 2.07 ms per loop
# master --> 100 loops, best of 3: 2.13 ms per loop
In [29]: %timeit sb.to_pydatetime()
# PR     --> 100 loops, best of 3: 2.08 ms per loop
# master --> 100 loops, best of 3: 2.12 ms per loop
```

Trying a method that will go through `get_start_end_field`.  The `BH` versions actually go through the changed weekday code paths:

```
In [30]: %timeit s.is_year_end
# PR     --> 1000 loops, best of 3: 307 µs per loop
# master --> 1000 loops, best of 3: 1.7 ms per loop

In [31]: %timeit sb.is_year_end
# PR     --> 1000 loops, best of 3: 362 µs per loop
# master --> 100 loops, best of 3: 1.78 ms per loop

In [32]: %timeit sb.is_quarter_start
# PR     --> 1000 loops, best of 3: 345 µs per loop
# master --> 100 loops, best of 3: 1.73 ms per loop
```


> Given that asv is so noisy on your laptop,

BTW, this is from a fairly beefy desktop in my basement.  There is non-zero background work running that can account for some noise, but both the level and variation should be much smaller than a laptop.